feat: add LocationManager service and organize folder structure
