+++ Wandering/Wandering/Features/Map/Service/LocationManager.swift
@@ -1,0 +1,61 @@
+//
+//  LocationManager.swift
+//  Wandering
+//
+//  Created by 이유정 on 6/26/25.
+//
+
+import Foundation
+import CoreLocation
+import MapKit
+
+class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+    private var manager = CLLocationManager()
+
+    @Published var userLocation: CLLocation?
+    @Published var isAuthorized = false
+    @Published var authorizationStatus: CLAuthorizationStatus = .notDetermined
+    @Published var region = MKCoordinateRegion(
+        center: CLLocationCoordinate2D(latitude: 51.5074, longitude: -0.1278),
+        span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
+    )
+
+    override init() {
+        super.init()
+        manager.delegate = self
+        manager.desiredAccuracy = kCLLocationAccuracyBest
+        manager.requestWhenInUseAuthorization()
+    }
+
+    func requestLocationPermission() {
+        manager.requestWhenInUseAuthorization()
+    }
+
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        authorizationStatus = manager.authorizationStatus
+
+        switch manager.authorizationStatus {
+        case .authorizedWhenInUse, .authorizedAlways:
+            isAuthorized = true
+            manager.startUpdatingLocation()
+        case .denied, .restricted:
+            isAuthorized = false
+        case .notDetermined:
+            manager.requestWhenInUseAuthorization()
+        @unknown default:
+            break
+        }
+    }
+
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        guard let location = locations.last else { return }
+        userLocation = location
+
+        DispatchQueue.main.async {
+            self.region = MKCoordinateRegion(
+                center: location.coordinate,
+                span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
+            )
+        }
+    }
+}
