+++ Wandering/Wandering/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
