+++ Wandering/Wandering/Resources/Assets.xcassets/google.imageset/Contents.json
